Switch to a more permissive verification check.
EIP-3074 has two versions, which I'll refer to as "strong" and "weak".
In the strong version, which is the currently specified version,
`tx.origin` is allowed to equal `msg.sender` after an `AUTHCALL`. This
is the preferred version of the EIP, but does possibly introduce some
minor problems for exceeding rare cases.

The "weak" version, on the other hand, does not allow `tx.origin` to
equal `msg.sender`.

This commit will allow the "weak" version of EIP-3074 to unlock the
bribes.